+++ src/types.ts
@@ -8,6 +8,7 @@ export type MapSettings = {
 };
 
 export type MarkerAction = (marker: ymaps.GeoObject) => void;
+export type UpdateFunction = (arr: ymaps.GeoObject[], action: 'add' | 'remove') => void;
 export type MarkerType = 'Point' | 'LineString' | 'Rectangle' | 'Polygon' | 'Circle';
 export interface Geometry extends ymaps.IGeometryJson {
   coordinates: Array<RecursiveArray | number>;
@@ -17,5 +18,6 @@ export interface MarkerFeature {
   geometry: Geometry;
   properties: object;
 }
+
 export type RecursiveArray = Array<RecursiveArray | number>;
 export type MapType = 'yandex#map' | 'yandex#satellite' | 'yandex#hybrid' | undefined;
