+++ src/use/marker-actions.ts
@@ -1,8 +1,6 @@
-import { MarkerAction } from './types';
+import { MarkerAction, UpdateFunction } from '../types';
 
-type UpdateFunction = (arr: ymaps.GeoObject[], action: 'add' | 'remove') => void;
-
-export default function useGeoObjectActions(updateFn: UpdateFunction) {
+export default function useGeoObjectActions(updateFunction: UpdateFunction) {
   const markersToAdd: ymaps.GeoObject[] = [];
   const markersToDelete: ymaps.GeoObject[] = [];
   let addGeoObjectTimeout: number | undefined;
@@ -11,13 +9,13 @@ export default function useGeoObjectActions(updateFn: UpdateFunction) {
   const addGeoObject: MarkerAction = (marker: ymaps.GeoObject) => {
     markersToAdd.push(marker);
     clearTimeout(addGeoObjectTimeout);
-    addGeoObjectTimeout = window.setTimeout(() => updateFn(markersToAdd, 'add'));
+    addGeoObjectTimeout = window.setTimeout(() => updateFunction(markersToAdd, 'add'));
   };
 
   const deleteGeoObject: MarkerAction = (marker: ymaps.GeoObject) => {
     markersToDelete.push(marker);
     clearTimeout(deleteGeoObjectTimeout);
-    deleteGeoObjectTimeout = window.setTimeout(() => updateFn(markersToDelete, 'remove'));
+    deleteGeoObjectTimeout = window.setTimeout(() => updateFunction(markersToDelete, 'remove'));
   };
 
   return { addGeoObject, deleteGeoObject };
